+++ CHANGELOG.rst
@@ -9,12 +9,17 @@ This project adheres to `Semantic Versioning`_ starting with version 0.2.0.
 
 Added
 -----
-
 - added scipy dependency
+- added element representation for command-line output
+
+Changed
+-------
+- improved button representation for custom buttons in command-line
 
 Removed
 -------
 - removed keras dependency, since keras_policy uses tf.keras
+
 
 [0.12.2] - 2018-11-20
 ^^^^^^^^^^^^^^^^^^^^^
@@ -23,6 +28,7 @@ Fixed
 -----
 - argument handling on evaluate script
 - added basic sanitization during visualization
+
 
 [0.12.1] - 2018-11-11
 ^^^^^^^^^^^^^^^^^^^^^